restore basic functionality displaying weather icons
As written in the documentation of Weathericon, the icons are not meant to be called on demand like
earlier versions, but to be downloaded and packed with your application. Therefore add the
repository containing the weather icons and use the image files for display.
Moreover, delete code for on-the-fly image manipulation inside Emacs with ImageMagick.